requirements-dev.txt
pynacl
python-dotenv
aws_lambda_powertools
